+++ src/Albatross.Expression/Operations/Year.cs
@@ -20,11 +20,11 @@ namespace Albatross.Expression.Operations {
 			object value = Operands.First().EvalValue(context);
 			if (value == null) { return null; }
 			else if (value is DateTime) {
-				return Convert.ToDouble(((DateTime)value).Year);
+				return Convert.ToDecimal(((DateTime)value).Year);
 			} else {
 				DateTime datetime;
 				if (DateTime.TryParse(Convert.ToString(value), out datetime)) {
-					return Convert.ToDouble(datetime.Year);
+					return Convert.ToDecimal(datetime.Year);
 				} else {
 					throw new FormatException("Invalid DateTime Format");
 				}
